move unified clone logic into clone.py
